[XrdCl] Parse correctly CDFH ZIP64 extension.

+++ src/XrdCl/XrdClZipArchiveReader.hh
@@ -144,7 +144,7 @@ class ZipArchiveReader
     //!
     //! @return         : the size of the file as in CDFH record
     //------------------------------------------------------------------------
-    XRootDStatus GetSize( const std::string &filename, uint32_t &size ) const;
+    XRootDStatus GetSize( const std::string &filename, uint64_t &size ) const;
 
     //------------------------------------------------------------------------
     //! Check if the archive is open
